Keep Project data aside
