+++ app/projects/page.tsx
@@ -3,34 +3,7 @@ import Head from "next/head";
 import Navbar from "../../components/Navbar";
 import ProjectCard from "../../components/ProjectCard";
 import "../../styles/projects.css";
-
-const projects = [
-  {
-    title: "Neural Interface Dashboard",
-    description:
-      "An immersive 3D dashboard that visualizes complex data streams through an intuitive neural interface. Features interactive 3D elements and real-time data processing.",
-    techStack: ["Next.js", "Three.js", "WebGL", "Node.js", "TensorFlow.js"],
-    image: "/api/placeholder/600/400",
-    tag: "Featured",
-    links: [
-      { label: "Live Demo", href: "#" },
-      { label: "Source Code", href: "#" },
-    ],
-  },
-  {
-    title: "CyberChat LLM",
-    description:
-      "A cyberpunk-themed chat application powered by AI language models with custom UI and animations.",
-    techStack: ["React", "Node.js", "OpenAI API", "Three.js"],
-    image: "/api/placeholder/400/250",
-    tag: "Web App",
-    links: [
-      { label: "Demo", href: "#" },
-      { label: "GitHub", href: "#" },
-    ],
-  },
-  // Add other projects here...
-];
+import projects from "./projectsData";
 
 export default function ProjectsPage() {
   return (
